Extracted the MovieList component logic into a new useMovieList custom hook
