+++ src/features/MovieList/index.js
@@ -1,18 +1,4 @@
-import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useHistory } from "react-router-dom";
-import {
-  fetchMoviesListLoad,
-  selectMoviesListState,
-  selectMoviesList,
-  selectStatus,
-  selectGenres
-} from "./movieListSlice";
-import {
-  selectPage,
-  resetPage
-} from "../../common/Pagination/paginationSlice";
-import { useQueryParameter } from "../../common/NavigationBar/Search/queryParameter";
+import { useMovieList } from "./useMovieList";
 import MovieTile from "../../common/Tiles/MovieTile";
 import Pagination from "../../common/Pagination";
 import Loading from "../Actions/Loading";
@@ -22,38 +8,15 @@ import { MovieListOrganizer } from "./styled";
 import { SectionHeading } from "../../common/styled";
 
 const MovieList = () => {
-  const dispatch = useDispatch();
-  const history = useHistory();
+  const {
+    status,
+    moviesList,
+    movieGenres,
+    totalPages,
+    currentPage,
+    query
+  } = useMovieList();
 
-  const status = useSelector(selectStatus);
-  const moviesList = useSelector(selectMoviesList);
-  const movieGenres = useSelector(selectGenres);
-
-  const totalPages = 500;
-  const { currentPage } = useSelector(selectMoviesListState);
-  const selectedPage = useSelector(selectPage);
-
-  useEffect(() => {
-    const unlisten = history.listen(() => {
-      dispatch(resetPage());
-    });
-
-    return () => {
-      unlisten();
-    };
-  }, [dispatch, history]);
-
-  useEffect(() => {
-    dispatch(fetchMoviesListLoad({ page: currentPage }));
-  }, [dispatch, currentPage]);
-
-  useEffect(() => {
-    if (currentPage !== selectedPage) {
-      dispatch(fetchMoviesListLoad({ page: selectedPage }));
-    }
-  }, [dispatch, currentPage, selectedPage]);
-
-  const query = useQueryParameter("search");
   if (query) {
     return <SearchMovies movieGenres={movieGenres} />;
   }
